+++ src/Runtime/StockFlowTransformer.Utilities.cs
@@ -74,7 +74,7 @@ namespace SyncroSim.STSimStockFlow
                         SpatialOutputFlowRecord rec = new SpatialOutputFlowRecord();
 
                         rec.FlowTypeId = ft.Id;
-                        rec.Data = new double[this.STSimTransformer.Cells.Count];
+                        rec.Data = new float[this.STSimTransformer.Cells.Count];
                         rec.HasOutputData = false;
 
                         this.m_SpatialOutputFlowDict.Add(ft.Id, rec);
@@ -107,7 +107,7 @@ namespace SyncroSim.STSimStockFlow
                         SpatialOutputFlowRecord rec = new SpatialOutputFlowRecord();
 
                         rec.FlowTypeId = ft.Id;
-                        rec.Data = new double[this.STSimTransformer.Cells.Count];
+                        rec.Data = new float[this.STSimTransformer.Cells.Count];
                         rec.HasOutputData = false;
 
                         this.m_LateralOutputFlowDict.Add(ft.Id, rec);
